like this?
